### 2024-12-03 #7563 @editorialbot
[REVIEW]: pridepy: A Python package to download and search data from PRIDE database
**Submitting author:** @ypriverol (<a href="http://orcid.org/0000-0001-6579-6941">Yasset Perez-Riverol</a>)
**Repository:** https://github.com/PRIDE-Archive/pridepy
**Branch with paper.md** (empty if default branch): master
**Version:** v0.0.8
