### 2024-12-06 PRIDE-Archive/pridepy#46 @cramacha
JOSS Review Comments
Hello,

I am one of the reviewers tasked with reviewing your submission for JOSS (See issue: https://github.com/openjournals/joss-reviews/issues/7563) and I have a minor list of comments that could be addressed to improve the quality of the project.

1. In the README file, you don't mention the usage of `get-files-by-project-accession`, `get-private-files`, `get-projects`. It would be nice to add details about those.
2. In the README, it would be useful to mention how one can contribute to the repo.
3. In the paper, you have done an exercise of performing a benchmark of download speeds for different protocols. Did you do any exercise comparing your tool with any similar tool?
4. Is there any documented list of future enhancements that could be done for the tool? For example, maybe you could improve download speeds or perhaps implement a GUI for managing clients and downloads.
5. I am being nitpicky here, but in your LICENSE file, I think you may need to replace the generic "Copyright [yyyy] [name of copyright owner]" with the actual owners?
6. It would be useful to add more comments on these files: https://github.com/PRIDE-Archive/pridepy/blob/master/pridepy/pridepy.py explaining the purpose and overall usage. Another place is here: https://github.com/PRIDE-Archive/pridepy/blob/master/pridepy/project/project.py. You could add missing comments as needed.

Otherwise, thanks for the opportunity to review this code.
