### 2024-12-07 PRIDE-Archive/pridepy#47 @MohmedSoudy
JOSS Review comments
Thank you for submitting your paper to The Journal Of Open Software (JOSS)  (Issue: openjournals/joss-reviews#7563). I have some minor comments that might enhance the paper.

- The statement of need should be enhanced by adding the limitation and comparing it with other available packages that do the same functionality, such as ppx in Python and prideR in R.
- The benchmarking of download speed should be modified with the packages mentioned above to show the differences between them and your package.
- In GitHub examples, It would be great if you could add more examples about different filters that could be applied by the user.
