### 2024-12-07 #7563 @MohmedSoudy
- [x] **Repository:** Is the source code for this software available at the [pridepy]( https://github.com/PRIDE-Archive/pridepy)?
- [x] **Contribution and authorship:** Has the submitting author [ypriverol](https://github.com/ypriverol) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
- [x] **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participant's research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
