### 2024-12-13 PRIDE-Archive/pridepy#47 @ypriverol
Hi @MohmedSoudy thanks for reviewing the library, here are some comments:

- prideR (https://github.com/PRIDE-R/prideR) was a library originally developed to interact between R and PRIDE API 2014. The library is not maintained anymore, mainly because the majority of the users use Python to download and interact with the API. prideR hasn't been maintained anymore since 2015 (more than 10 years ago), so it would be difficult to benchmark.
- ppx aims to retrieve data from multiple ProteomeXchange databases using https, and it is maintained by a team outside PRIDE Archive. pridepy is maintained by the PRIDE team, and it is only dedicated to PRIDE resources, it supports multiple PRIDE protocols Aspera, HTTP (as ppx) and Globus HTTPs, it also supports downloads from private datasets. I do think is good to acknowledge in the manuscript the existence of ppx. We will modify the manuscript.
- We will add more examples about how to use the library.
