### 2025-01-08 PRIDE-Archive/pridepy#46 @selvaebi
Hi @cramacha,

Thank you for your valuable comments! Based on your suggestions, we have created a pull request:

🔗 https://github.com/PRIDE-Archive/pridepy/pull/48

Here are a few points for clarification:

Python CLI Tool: This tool, built using the Click module, already provides detailed usage instructions for each command. These instructions are directly accessible from the CLI, ensuring clarity and avoiding redundancy in the README.
Code Comments: We believe the existing comments in the files are sufficient to explain the functionality.
Comparison Tools: Since this is already a command-line tool using different protocols, we didn't use any other tools for comparison.
Future Enhancements: At this moment, we don't have any planned enhancements, but we are always open to ideas.
Thanks again for your feedback!
