### 2025-01-08 PRIDE-Archive/pridepy#46 @ypriverol
@cramacha I will close this issue if you are happy with the changes. or you can close it yourself?
