+++ recipe/meta.yaml
@@ -1,7 +1,7 @@
 # recipe/meta.yaml
 package:
   name: pridepy
-  version: "0.0.3"
+  version: "0.0.6"
 
 source:
   path: ../
