### 2025-01-08 #7563 @ypriverol
Dear @cramacha @MohmedSoudy @jedbrown :

First thanks to both reviewers for taking the time to review our code and library. In these PRs (https://github.com/PRIDE-Archive/pridepy/pull/48) and (https://github.com/PRIDE-Archive/pridepy/pull/49) we have tried to tackle all the issues and comments as requested by the reviewers. One of the major comments is the lack of comparison with two existing tools:
- prideR and ppx. prideR was a previous package (not under maintenance anymore) that allowed us to interact with PRIDE API v1. This package is not maintained anymore and would be difficult to benchmark with pridepy, as it is not properly working.
- ppx is a general-purpose library that enables the download of data from multiple ProteomeXchange databases using mainly HTTP. pridepy is more than that, first, it allows to download using HTTP, FTP, and Globus-HTTP, and it allows to download private and public datasets and files. In addition, pridepy allows interaction with PRIDE API, enabling the filtering of files in the API by file type and interacting with the latest version of PRIDE API. Nevertheless, we updated the manuscript mentioning ppx and what new features pridepy allows that ppx does not include.

During the review process, we added more examples to the README and updated the LICENSE, manuscript, code and way How to contribute to the package.

Please let us know if you need to include more changes before the new release of the library version 0.0.6 including all changes.
