### 2025-01-16 #7563 @ypriverol
For me, it is ready from our side. I did a small update to the PDF about the funding, and now it is ready.
