### 2025-02-06 #7563 @danielskatz
@ypriverol - (just jumping in for a second, as I happened to see this).  You have to do the "author" steps in https://github.com/openjournals/joss-reviews/issues/7563#issuecomment-2597077431 before anything else can happen
