### 2025-02-06 #7563 @danielskatz
As stated above, you need to

* Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
* Archive the release on Zenodo/figshare/etc and post the DOI here.
* Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
* Make sure that the license listed for the archive is the same as the software license.
