+++ pridepy/files/files.py
@@ -51,8 +51,8 @@ class Files:
     """
 
     V3_API_BASE_URL = "https://www.ebi.ac.uk/pride/ws/archive/v3"
-    API_BASE_URL = "https://www.ebi.ac.uk/pride/ws/archive/v2"
-    API_PRIVATE_URL = "https://www.ebi.ac.uk/pride/private/ws/archive/v2"
+    API_BASE_URL = "https://www.ebi.ac.uk/pride/ws/archive/v3"
+    API_PRIVATE_URL = "https://www.ebi.ac.uk/pride/private/ws/archive/v3"
     PRIDE_ARCHIVE_FTP = "ftp.pride.ebi.ac.uk"
     S3_URL = "https://hh.fire.sdo.ebi.ac.uk"
     S3_BUCKET = "pride-public"
