### 2025-02-07 ad6faf1
fix: removed dataegap and test searchby keyword
