### 2025-02-08 3c156d2
Update pridepy/tests/test_search.py
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
