+++ README.md
@@ -31,7 +31,7 @@ Install with setup.py:
 ```bash
 $ git clone https://github.com/PRIDE-Archive/pridepy
 $ cd pridepy
-$ python project.toml sdist bdist_wheel 
+$ python pyproject.toml sdist bdist_wheel 
 $ pip install dist/pridepy-{version}.tar.gz
 ```
 
