+++ pyproject.toml
@@ -1,41 +1,109 @@
-[build-system]
-requires = ["setuptools>=45", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
+[tool.poetry]
 name = "pridepy"
 version = "0.0.7"
 description = "Python Client library for PRIDE Rest API"
 authors = [
-    {name = "PRIDE Team", email = "pride-support@ebi.ac.uk"},
+    "PRIDE Team <pride-support@ebi.ac.uk>",
+]
+maintainers = [
+    "PRIDE Team <pride-support@ebi.ac.uk>",
 ]
 readme = "README.md"
-requires-python = ">=3.7"
-keywords = ["PRIDE", "python", "client", "REST", "API"]
 license = "Apache-2.0"
-dependencies = [
-    "requests",
-    "ratelimit",
-    "click",
-    "pytest",
-    "setuptools",
-    "plotly",
-    "tqdm",
-    "boto3",
-    "botocore",
-    "httpx",
+keywords = [
+    "PRIDE",
+    "proteomics",
+    "mass-spectrometry",
+    "API",
+    "big data",
+    "bioinformatics",
+    "data science",
+]
+classifiers = [
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent",
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.7",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Topic :: Scientific/Engineering :: Bio-Informatics",
+    "Development Status :: 5 - Production/Stable"
+]
+packages = [
+    { include = "pridepy" }
+]
+include = [
+    "LICENSE",
+    "README.md",
+    "**/*.txt",
+    "**/*.md",
+    "aspera/**/*"
 ]
 
-[project.urls]
-Homepage = "https://github.com/PRIDE-Archive/pridepy"
+[tool.poetry.dependencies]
+python = "^3.7"
+requests = "^2.31.0"
+ratelimit = "^2.2.1"
+click = "^8.1.7"
+plotly = "^5.18.0"
+tqdm = "^4.66.1"
+boto3 = "^1.34.0"
+botocore = "^1.34.0"
+httpx = "^0.27.0"
 
-[project.scripts]
+[tool.poetry.group.dev.dependencies]
+pytest = "^8.0.0"
+black = "^24.1.0"
+isort = "^5.13.0"
+pytest-cov = "^4.1.0"
+mypy = "^1.8.0"
+
+[tool.poetry.urls]
+PyPI = "https://pypi.org/project/pridepy/"
+PRIDE = "https://www.ebi.ac.uk/pride/archive/"
+Documentation = "https://github.com/PRIDE-Archive/pridepy#readme"
+Repository = "https://github.com/PRIDE-Archive/pridepy"
+License = "https://github.com/PRIDE-Archive/pridepy/LICENSE"
+
+[tool.poetry.scripts]
 pridepy = "pridepy:main"
 
-[tool.setuptools]
-packages = ["pridepy"]
-py-modules = ["pridepy"]
+[tool.black]
+line-length = 99
+target-version = ["py39"]
+include = '\.pyi?$'
+extend-exclude = '''
+# A regex preceded with ^/ will apply only to files and directories
+# in the root of the project.
+^/tests/
+'''
 
-[tool.setuptools.package-data]
-"*" = ["*.txt", "*.md"]
-"aspera" = ["**/*"]
+[tool.isort]
+profile = "black"
+multi_line_output = 3
+line_length = 99
[…]
