### 2025-02-09 fd355b3
fix: space in fasta
