### 2025-02-10 1d95677
Update pridepy/pridepy.py
Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>
