+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pridepy"
-version = "0.0.7"
+version = "0.0.8"
 description = "Python Client library for PRIDE Rest API"
 authors = [
     "PRIDE Team <pride-support@ebi.ac.uk>",
