### 2025-03-07 #7563 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
