### 2025-03-10 #7563 @danielskatz
The track editor (@openjournals/bcm-eics) will handle the next steps.
