### 2025-03-17 #7563 @Kevin-Mattheus-Moerman
@ypriverol I'll update the paper now, but there is no need to update the ZENODO archive (unless you want to).
